refactor: move utils into WalletResponse; organise types
