+++ packages/server-wallet/src/wallet/multi-threaded-wallet/index.ts
@@ -1,7 +1,8 @@
 import {UpdateChannelParams} from '@statechannels/client-api-schema';
 
 import {IncomingServerWalletConfig} from '../../config';
-import {MultipleChannelOutput, SingleChannelOutput, SingleThreadedWallet} from '../wallet';
+import {MultipleChannelOutput, SingleChannelOutput} from '../types';
+import {SingleThreadedWallet} from '../wallet';
 
 import {WorkerManager} from './manager';
